+++ HPXMLtoOpenStudio/tests/test_enclosure.rb
@@ -688,41 +688,6 @@ class HPXMLtoOpenStudioEnclosureTest < MiniTest::Test
     _check_surface(hpxml.furniture_mass, os_surface, furniture_mass_layer_names)
   end
 
-  def test_compartmentaliztion_area
-    # Test single-family detached
-    hpxml = _create_hpxml('base.xml')
-    total_area, exterior_area = hpxml.compartmentalization_boundary_areas()
-    a_ext_ratio = exterior_area / total_area
-    assert_equal(1.0, a_ext_ratio)
-
-    hpxml = _create_hpxml('base-foundation-unconditioned-basement.xml')
-    total_area, exterior_area = hpxml.compartmentalization_boundary_areas()
-    a_ext_ratio = exterior_area / total_area
-    assert_equal(1.0, a_ext_ratio)
-
-    hpxml = _create_hpxml('base-atticroof-cathedral.xml')
-    total_area, exterior_area = hpxml.compartmentalization_boundary_areas()
-    a_ext_ratio = exterior_area / total_area
-    assert_equal(1.0, a_ext_ratio)
-
-    # Test single-family attached
-    hpxml = _create_hpxml('base-bldgtype-single-family-attached.xml')
-    total_area, exterior_area = hpxml.compartmentalization_boundary_areas()
-    a_ext_ratio = exterior_area / total_area
-    assert_in_delta(0.840, a_ext_ratio, 0.001)
-
-    hpxml.attics[0].within_infiltration_volume = true
-    total_area, exterior_area = hpxml.compartmentalization_boundary_areas()
-    a_ext_ratio = exterior_area / total_area
-    assert_in_delta(0.817, a_ext_ratio, 0.001)
-
-    # Test multifamily
-    hpxml = _create_hpxml('base-bldgtype-multifamily.xml')
-    total_area, exterior_area = hpxml.compartmentalization_boundary_areas()
-    a_ext_ratio = exterior_area / total_area
-    assert_in_delta(0.247, a_ext_ratio, 0.001)
-  end
-
   def test_kiva_initial_temperatures
     initial_temps = { 'base.xml' => 68.0, # foundation adjacent to conditioned space, IECC zone 5
                       'base-foundation-conditioned-crawlspace.xml' => 68.0, # foundation adjacent to conditioned space, IECC zone 5
@@ -833,7 +798,7 @@ class HPXMLtoOpenStudioEnclosureTest < MiniTest::Test
 
   def test_aspect_ratios
     # Test single-family attached
-    hpxml = _create_hpxml('base-bldgtype-single-family-attached.xml')
+    hpxml = _create_hpxml('base-bldgtype-attached.xml')
     wall_outside = hpxml.walls.select { |w| w.exterior_adjacent_to == HPXML::LocationOutside && w.interior_adjacent_to == HPXML::LocationLivingSpace }[0]
     wall_other_housing_unit = hpxml.walls.select { |w| w.exterior_adjacent_to == HPXML::LocationOtherHousingUnit && w.interior_adjacent_to == HPXML::LocationLivingSpace }[0]